Allow creating a admin `ClusterRoleBinding`
This PR adds the possibility of creating a cluster role binding between a group whose name defined by `adminClusterRolebinding.name` and the built-in `cluster-admin` `ClusterRole`.
This is particularly useful for GKE Autopilot clusters where it's not
possible to use the default `system:masters` group because authz Warden
security module prevents impersonating system-wide identities.

When the chart detects that the target cluster is a GKE cluster - `version: v1.x.x-gke.<build>` - it will automatically create the `ClusterRoleBinding` and print a warning message with the following payload:

```
NOTES:
WARNING: GKE Autopilot clusters forbid users from impersonating system-wide identities.
This means you won't be able to use the `system:masters` Kubernetes Group in
the Teleport Roles for GKE Autopilot clusters.

Given that you installed Teleport on a GKE cluster, we recommend you use the
Kubernetes Group `cluster-admin` instead of `system:masters` in the Teleport Roles
for GKE Autopilot clusters.

This chart automatically created the `cluster-admin` Kubernetes Group for you and
assigned it admin privileges on the Kubernetes cluster.

Consult the built-in security features that GKE Autopilot enforces:
https://cloud.google.com/kubernetes-engine/docs/concepts/autopilot-security#built-in-security
```

Part of #28506
Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>