Update the Documentation Test Plan (#29129)
Closes #27755

- Separate the docs test plan into its own issue template. This makes
  the checklist easier to work with for Docs Team members. Also adds
  sections to the Docs Test Plan for readability.

- Update the list of guides to verify. This reflects the current range
  of getting started guides available for Teleport's editions.

- Propose verifying that we can complete all how-to guides introduced
  within an upcoming release.

We need to strike a balance between testing critical guides and keeping
the workload of the Docs Team manageable, so this change limits the
number of guides it adds to the Test Plan.

Alongside new feature guides, this change adds the Terraform AWS HA
guide to the list of guides to verify. This is because we want to
position this as a critical guide for launching an HA self-hosted
Teleport cluster.