Switch to upstream x/crypto (#28929) (#29601)
* Switch to upstream x/crypto

Removed the dependency override for golang.org/x/crypto and updated the specific version used latest in go.mod. This change is needed because the specific issue with OpenSSH <= 7.6 that necessitated the override has been addressed upstream, making the override unnecessary.

* Update config test to align with newer encryption standards

Commit updates the config test to include more encryption standards, thus maintaining compatibility with a wider range of environments. "ElementsMatch" is now used to ensure the presence of cipher keys, irrespective of their order in the list. This makes the tests more resilient against future changes and amendments in encryption standards.