[v13] feat: Login Rule support for email.local and regexp.replace (#29611)
* feat: Login Rule support for email.local and regexp.replace

This commit updates the `stringListMap` function to filter out values
that map to the empty string.
This is used by `email.local` and `regexp.replace` only.
The empty values were already being filtered out in
`(*TraitsTemplateExpression).Interpolate`, adding the filtering here
just makes it more convenient to call these functions in the Login Rule
parser in teleport.e where it's desirable to have the same behaviour of
`regexp.replace` filtering out values that don't match the expression.

The actual implementation and tests for the new helpers in Login Rules
is added in a buddy PR in teleport.e.

This PR also adds documentation for the new functions.

* apply docs grammar suggestions

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>

---------

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>