Update session TTL handling and fix access request string representation
Modified the access request logic to consider the maxDuration flag while setting the sessionTTL. Instead of simply replacing sessionTTL with the maxDuration (when maxDuration is set), the minimum value of sessionTTL and maxDuration is now used. This change has been reflected in the unit test as well. Also, the string representation of AccessRequest is fixed to accurately represent the user and roles.

The changes improve the handling of sessionTTL, providing a more accurate session duration that considers both session TTL and max duration. Furthermore, this improves AccessRequest debugging by providing a more descriptive string representation.