[tctl] Adds option to write tarred `tctl auth sign` output to stdout (#29451)
* Proof-of-concept writing tar to stdout

A quick and dirty experiment showing one possible approach for
writing certificates to stdout. Demonstrates a possible solution
to #29262.

DO NOT MERGE AS IS. IN NO WAY PRODUCTION READY.

* Fix timestamps

* Adds `--tar` option to `auth sign`

Adds an option to bundle the certificates generated by `tctl auth sign`
into a tarball and writes that tarball to stdout.

This is to facilitate extracting credentials from environments with
limited access to the filesystem and tools like a shell, tar and so
on, e.g. distroless Docker images.

Example usage:

```
$ kubectl exec ... -- tctl auth sign --user alice --format openssh -o alice --tar | tar xv
x alice-cert.pub
x alice
```

* Reroutes helper mesg to stderr when straming tar file to stdout