build: Revert commits saving deb/rpm artifacts (#1890)
Revert the two commits that save the deb/rpm artifacts. Both of them
together work fine on master and branch/v13, but branch/v12 and
branch/v11 do not have a precusor commit required which has caused much
head scratching and weirdness on the v11 release.

Just revert these for now - they are not needed for a drone release.

Source-commit: gravitational/teleport.e@1cb8ec76c9052cf0865cb663124b603e4d72e2fd
Source-pull-requests: gravitational/teleport.e#1890