Update MongoDB Driver to master commit (#29656)
This updates the MongoDB driver to use a recent master commit in order to incorporate the fix: https://github.com/mongodb/mongo-go-driver/pull/1291 (which has not been released to a tag yet)

In addition a fuzz seed was added to cover this condition.  After not finding more issues this also re-enables the MongoDB fuzzing in oss-fuzz.