[v12] Hold Auth init lock for the duration of initialization (#29709)
* Add configuration with optional timeout to AcquireLock (#24559)

* Add configuration with optional timeout to AcquireLock

* rename to RetryInterval

* backport RunWhileLocked changes from #25639

* Hold Auth init lock for the duration of initialization (#29593)

Auth now uses `RunWhileLocked` instead of `AcquireLock` to ensure
that the initialization lock is held until the bootstrapping
process is completed. Prior, Auth only held the lock for 30s which
could allow multiple Auths to attempt bootstrapping simultaneously.
Initialization should complete prior to 30s in most cases, but it
is not guarateed, especially on first boot when CAs are being
generated and a large data migration may be needed.

---------

Co-authored-by: Tobiasz Heller <14020794+tobiaszheller@users.noreply.github.com>