Fix Kubernetes Legacy Proxy heartbeats
When a Teleport Kubernetes Proxy runs in legacy mode - i.e. - spec enables `proxy_service.kubernetes.enabled`, the proxy is capable of proxy requests to a target Kubernetes cluster instead of forwarding them to a Teleport Kubernetes Service.

The following config enables the legacy proxy mode in proxy:

```yaml
proxy_service:
    enabled: yes
    web_listen_addr: 0.0.0.0:3080
    public_addr: tele.local:3080
    kubernetes:
      enabled: yes
      listen_addr: 0.0.0.0:3026
      kubeconfig_file: path # optional
```

Because the proxy forwards and heartbeat the configured Kube clusters, it needs permissions to upsert the kube cluster in auth's backend and to extend its expiration date.

This PR fixes a problem introduced in Teleport 8 where the proxy is allowed to upsert the kube server, but it's not allowed to extend the expiration time.

When the request fails, the proxy becomes degraded and the performance is affected.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>