Add SessionTTL to the Access Request object (#29658)
* Add SessionTTL to the Access Request object

New field Session TTL holds information on how long a generated certificate will be valid for when the max duration is used. The newly implemented V2 API endpoint is to be used for this purpose, to ensure backwards compatibility. The V1 API usage marked for deprecation in version 15.0.0.

* Update session TTL handling and fix access request string representation

Modified the access request logic to consider the maxDuration flag while setting the sessionTTL. Instead of simply replacing sessionTTL with the maxDuration (when maxDuration is set), the minimum value of sessionTTL and maxDuration is now used. This change has been reflected in the unit test as well. Also, the string representation of AccessRequest is fixed to accurately represent the user and roles.

The changes improve the handling of sessionTTL, providing a more accurate session duration that considers both session TTL and max duration. Furthermore, this improves AccessRequest debugging by providing a more descriptive string representation.

* Deprecate old CreateAccessRequest method

The old CreateAccessRequest method is marked as deprecated in the proto files. This commit indicates it will be deleted in v15.0.0. Users are urged to use CreateAccessRequestV2 instead to create new access requests.

* Fix test

* Fix test