Add Teleport and Kube Cluster name tags to Kube Prom metrics
This PR adds `kube_cluster` and `cluster` tags to Kubernetes Access
prometheus metrics. This labels are unknown at the handler so we used a
`prometheus.Labels` maps carried in the request's `ctx` to be able to
fill the unknown labels so that they can be retrieved at metric
collection time.

Metrics are collected after the request is handled by the downstream
handlers so we leverage that to fill unknown labels.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>