Add max duration parameter to access requests (#1827)
* Add max duration parameter to access requests

This commit adds a new functionality for setting a maximum duration for access requests. This is achieved by adding a new field, "MaxDuration", in the access request go struct, which is populated from the incoming JSON request. The value is treated as a string representing a duration, and is parsed into a time.Duration type using a helper function, parseDuration, added in this commit. The parsed duration is then added to the current time to set the max duration of the access request. This allows users to specify for how long the access should be granted when creating a request which can be useful in various scenarios like temporary access for maintenance, audits etc. A text input field is also added in the UI for taking this parameter.

* Add display and handling of maximum request duration

This update focuses on capturing and displaying the maximum duration of access requests. Previously, the maximum duration was ignored in the UI, making it impossible for users to see or define how long they're requesting access for. Now the `maxDuration` is captured when creating a request, displayed in the request view, and appropriately passed to and handled in the backend. The ability to view and define the maximum duration offers better transparency and control over access requests duration.

The parsing functionality that previously was present but unused has consequently been cleaned up, which simplifies and streamlines the code.

* yarn prettier-write

* Update snaps

* Fix tests/linters

* Add SessionTTL to the Access Request object

New field Session TTL holds information on how long a generated certificate will be valid for when the max duration is used. The newly implemented V2 API endpoint is to be used for this purpose, to ensure backwards compatibility. The V1 API usage marked for deprecation in version 15.0.0.

* Update max access duration field and refactor code for duration handling

Replaced maxDurationDuration with maxDurationText for more clarity. Also added a new function (middleValues) to help display max access durations in a more human-friendly way. Some UI changes also made, including fields' positions and adding "Optional" to max duration time text. Refactored createAccessRequest and moved duration handling to makeDryRunRequest function. Tests and snapshots updated accordingly.

* "Refactor RequestView and tests to handle Date type durations"

This commit refactors the `<RequestView>` component and related tests to handle duration data as the `Date` type. Previously, durations were handled as strings which caused difficulties in correctly aggregating and comparing these values. The refactoring also includes introducing the `timeHelpers` module, intended for handling operations related to these new duration types. The changes in test files correspond to the updated ways of handling these durations.

This reformatting increases robustness of our time-related operations, increases overall code readability, and allows for a more intuitive handling of duration data types throughout the codebase.

* Address code review comments

* Add loading state for resource request checkout

This commit introduces a loading state when fetching resource request details in the RequestCheckout component. Previously, the component did not handle any loading state, which could lead to an incomplete rendering when a response was slow to return or failed. The state 'loading' and 'loaded' are introduced as type LoadingStatus. A loading spinner will now be displayed to inform the user that data is being fetched. If the fetch operation fails, it will still render the page without any max duration options.

* Fix code style issues in RequestCheckout and timeHelpers

This commit addresses minor code style issues. These changes include adding a semicolon to the constant SEVEN_DAYS_IN_MS in useRequestCheckout.tsx, replacing double quotes with single quotes in timeHelpers.test.ts and timeHelpers.ts for consistency, and also improving the readability in timeHelpers.ts by spreading imports into multiple lines. An extra space was added in a 'if' statement for clarity in the same file and a newline was added at the end of the file.

* Apply core review suggestions

* Apply core review suggestions

* Fix test compilation error

Source-commit: gravitational/teleport.e@cab33acc0961a589879931d72dbecf0687e30142
Source-pull-requests: gravitational/teleport.e#1827