[v13] dronegen: Build Teleport Connect for amd64 push build
Add an input parameter when calling the release-linux workflow to build
Teleport Connect for the AMD64 build. This was previously done when
Drone was doing the build but got accidentally dropped when moving to
GitHub actions.

This will also be used for the tag builds when they migrate to GHA as we
do a release build of Teleport Connect for each architecture.

Update .drone.yml with `make dronegen` to add the `build-connect`
parameter to the call of the `release-linux` workflow.

Backport: https://github.com/gravitational/teleport/pull/29622
Depends-on: https://github.com/gravitational/teleport.e/pull/1926
Changelog: none