dronegen: Build Teleport Connect for amd64 push build (#29622)
* dronegen: Build Teleport Connect for amd64 push build

Add an input parameter when calling the release-linux workflow to build
Teleport Connect for the AMD64 build. This was previously done when
Drone was doing the build but got accidentally dropped when moving to
GitHub actions.

This will also be used for the tag builds when they migrate to GHA as we
do a release build of Teleport Connect for each architecture.

* ci: Update .drone.yml

Update .drone.yml with `make dronegen` to add the `build-connect`
parameter to the call of the `release-linux` workflow.

* Update e ref for release-linux teleport param

Update e ref for 98fc02c3f276054b72fe7c55544b45834d964b9b so we can call
the release-linux workflow with the `build-connect` parameter.