kube: Implement full namespace jail
This PR implements a feature where it's possible to restrict access to
namespace scoped `CustomResources`.

To be able to deserialize unknown Kubernetes objects, we keep a
`serializer.CodecFactory` per cluster where we register not only the default
types - such as `Pods`, `Statefulsets`... -, but also all the
`CustomResourceDefinitions` that are namespaced.

With the list of RBAC supported resources and the serializer, Kube
access identifies if we support Kubernetes per-Resource RBAC for that
type - the match happens based on resource `kind` and `apiGroup` -
and enforces the RBAC rules.

Namespace controls access to `CustomResources`. When a user as an allow
rule for a namespace - similar to the rule bellow -, he has access to
all resources within that namespace including the custom resources.

```yaml
kubernetes_resources:
- kind: namespace
  name: "someNamespace"
  verbs: ['*']
```

The list of `CustomResources` is constantly updated and monitored in
intervals of `5m`. If a new CRD is created between two iteration cycles,
Teleport will only enforce RBAC after the resource and codecs are
updated.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>