Replace custom test renderer with gotestsum
All instances of the custom test renderer (`${RENDER_TESTS}`) in the Makefile have been replaced with gotestsum calls. This allows to provide a more standardized and extendible tool for handling the test output. The gotestsum tool, being a popular and actively maintained project, adds additional features such as output formatting options and compatibility with developer tools. As a result, it improves the readability and accessibility of our test results.