Replace custom test renderer with gotestsum (#29862)
* Replace custom test renderer with gotestsum

All instances of the custom test renderer (`${RENDER_TESTS}`) in the Makefile have been replaced with gotestsum calls. This allows to provide a more standardized and extendible tool for handling the test output. The gotestsum tool, being a popular and actively maintained project, adds additional features such as output formatting options and compatibility with developer tools. As a result, it improves the readability and accessibility of our test results.

* Add gotestsum and update Makefile, Dockerfile, and devbox files

Changes have been made in Makefile, Dockerfile, and devbox.json files to install gotestsum if it's not already installed and use it in the test commands. The devbox.lock file has also been updated accordingly to reflect the addition of gotestsum.

* Update gotestsum installation in Makefile

The Makefile has been adjusted to ignore any errors while installing 'gotestsum'. This is a temporary measure until 'gotestsum' is added to the Dockerfile, to prevent the build process from failing.

* Fix devbox?

* Apply a review suggestion
Revert devbox changes - devbox seems to be broken. I'll introduce this change in a new PR, so the current one is not blocked on an admin approval.

* Revert `make e2e-aws` to use render-test to prevent CI failing