dronegen: Convert some linux tag pipelines to GitHub Actions
Convert some of the linux-based tag build pipelines to run on GitHub
Actions. The following pipelines have been converted:

    build-linux-amd64
    build-linux-amd64-centos7
    build-linux-386
    build-linux-arm
    build-linux-amd64-deb
    build-linux-amd64-centos7-rpm
    build-linux-386-deb
    build-linux-386-rpm
    build-linux-arm-deb
    build-linux-arm-rpm

The GHA workflows builds tarballs as well as deb/rpm packages in the one
workflow, so the `-deb` and `-rpm` pipelines will need to be manually
removed from `.drone.yml`.

Still remaining of the linux-based tag build pipelines are the fips
pipelines which will need some special handing in the GitHub workflow,
the arm64 pipelines which are already converted using a different
workflow and the non-native windows build.