[v13] ci: Fix post-release calling update-ami-ids (#29886)
* [v13] ci: Fix post-release calling update-ami-ids

Fix the `post-release` workflow when calling the `update-ami-ids`
workflow. Two problems fixed:
1) The file extension is `.yaml` not `.yml`
2) You can't call a workflow in the current repo unless you check out
   the repo or you qualify the workflow with an owner/repository and a
   ref.

I've chosen to use the qualified form so we don't need to check out the
teleport repo just to call the `update-ami-ids` workflow as the previous
job already checked it out and `update-ami-ids` will also check it out,
so I feel a third checkout of the same repo is a bit excessive. Since
the workflow always runs of `master`, using the qualified form that
requires a ref works ok.

While we're here, fix up a couple of other things:

* `${{ vars.GITHUB_REF }}` is not correct. It was probably meant to be
  `${{ env.GITHUB_REF}}` (vars are from GitHub environments, not
  environment variables - different things), but I went with
  `${{ github.ref }}` as that is simpler.
* Quote a bunch of var expansions in the shell fragment, as actionlint
  and shellcheck complained about them.
* Use `gh`'s `-q` flag to remove the `jq | tr` pipeline.

This is hard to test since it triggers off the latest GitHub release.
Currently this workflow is not working so no harm if this is still not
right - we'll continue to tweak it.

* [v13] ci: Add missing dot to `gh -q` arg

I missed a dot in bc491d73b11aa31c88b59e5ed9805f433e8102dd in the `-q`
arg to `gh`. `--json` takes a field without the dot, `-q` needs the dot
(it's a `jq` expression).

Fixes: bc491d73b11aa31c88b59e5ed9805f433e8102dd