Fix Memory Aliasing in for loops (#30146)
This commit is a similar fix to 1022fe926c16629214177eafc412b32145a63b52

This fixes other discovered cases of referencing a for loop variable pointer.  Since the value is updated without the location changing it can result in only the last value being seen.
This is typically fixed by referencing the slice offset directly, or using a new variable.