EC2 Instance Connect Endpoint: HTTP endpoint to create Nodes (#29370) (#30189)
* EC2 Instance Connect Endpoint: HTTP endpoint to create Nodes

When the user receives a list of EC2 instances, they must be able to
pick one and add it to the cluster.

To do so, we must provide an API that accepts the node information and
adds it to the backend.

Only Nodes of type SubKindOpenSSHEC2InstanceConnectEndpointNode are
allowed.
Other nodes have their own join method which is not part of the HTTP
API.

* review pt1

* rename subkind

* fix import order