Fix a bug in infinite scroll.
This bug caused the UI to thrash with requests if the server returned an error
on the initial request. Note: this is a very crude fix, we need to refactor
this and do it properly. The downside of this implementation is that it's gonna
prevent the user from ever retrying the failed request and scroll down further.