Fix resources being deleted from Firestore on update (#30287)
The change to use `Update` in #28473 caused a 0 value to be written
to the expires field of the document. When using `Create` or `Set`
a zero expiry value would result in an empty expires field for that
document. Now that the expires column was containing a 0 value the
purgeExpiredDocuments loop was detecting the document as needing
to be removed. To prevent removing resources that have no expiry
the purge loop now explicitly ignores any resources that have a 0
expiry value.

Fixes #30236