ci: Use checkout-teleport action in kube-agent-updater (#1973)
Change the `release-teleport-kube-agent-updater-oci` workflow to use the
`checkout-teleport` action in this repository instead of the
`actions/checkout@v3` action, as the former can use a GitHub
application's token to check out the teleport OSS repository. This is
needed to be able to checkout teleport-private as the default token
perms on the teleport.e repo do not allow access to teleport-private.

Source-commit: gravitational/teleport.e@fb859f68881f7503f0107716926743f37d19f856
Source-pull-requests: gravitational/teleport.e#1973