[v13] EC2 Instance Connect Endpoint: List EC2 Instances (#30258)
* EC2 Instance Connect Endpoint: List EC2 Instances

We must list the EC2 instances to the user, so they can pick one and
then add them as a Node.

This PR adds EC2 Instance Listing.
To do this, we convert the EC2 Instance into Teleport Server with a
special subkind: SubKindOpenSSHEC2InstanceConnectEndpointNode

This is similar to the ListDatabases endpoint.

* review pt1

* fix const and add test

* use any instead of interface{}

* fix import name