Fix data race when calling Uploader's `Close` and `Serve` at the same time
This PR fixes a data race that was introduced by #30211 while trying to
fix the uploader write after being closed.

The data race happens when `Serve` and `Close` methods are called at the
same time - i.e. `Serve` is spawned by `RegisterFunc` and `Close` is
called via `OnExit` function.
There is no guarentee that when `OnExit` is called `Serve` is already
spawn and running. This creates a data race when `Close` blocks on
`wg.Wait()` but `Serve` still tries to call `wg.Add(1)` into the
waitgroup.

To prevent the data race, we added a `isClosing` field that controls
when the server is marked to be closing and we call `wg.Add(1)` if
`isClosing=false` and under locking to prevent the race between `Close`
and `Serve`.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>