Add a link to Teleport Labs in the landing page (#30382)
Fixes #29955

Adds this link in the Prerequisites section after the link to the
Teleport Team docs. This is a natural place for this information, since
the Teleport Team link is for users whose environments do not meet the
requirements of the Prerequisites of the Getting Started tutorial.

This change means that the `Notice` that includes the Teleport Team link
is now two paragraphs long, which is too long for a `Notice`. An
`Admonition`, however, would clash with the `Admonition` telling the
reader about our DigitalOcean 1-click droplet. I have removed the
colored boxes from the Prerequisites section to reduce noise.