Skip fatal error if `kube_cluster` is not available (#30433)
* Skip fatal error if `kube_cluster` is not available

This PR skips a fatal error if the `kube_cluster` isn't available when the
service starts and marks the cluster as offline if no schema can be
generated from the cluster. Offline clusters cannot serve any requests
but are still registered in Teleport. It spins a goroutine that tries to
generate a shema each 5min and if it's successfull, it marks the cluster
as online.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* fix error message to client

* handle isaiah code review

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>