Refactor AgentKind to UnifiedResource - build fix (#2009)
The 'AgentKind' was replaced with 'UnifiedResource' in 'useNewRequest.ts'. This was done to ensure more consistent type usage across the application. The 'kind' property was added to several objects in test files and 'user group' was added as a type to 'UnifiedResource' to accommodate this change.

Source-commit: gravitational/teleport.e@281a8677862a4b9c418519a02eacc3bed0c7fd0e
Source-pull-requests: gravitational/teleport.e#2009