Add extraVolumeMounts to the Operator sidecar container
When we start the sidecar contiainer, it seems to require access to the
files referenced in the config file.

```
2023-08-08T10:47:12Z	ERROR	setup	failed to connect to teleport cluster, backing off	{"error": "failed to create auth client config\n\topen /etc/teleport-join-tokens/kube-agent-dev-join-token: no such file or directory"}
main.main.func1
	/go/src/github.com/gravitational/teleport/integrations/operator/main.go:128
github.com/gravitational/teleport/api/utils/retryutils.(*Linear).For
	/go/src/github.com/gravitational/teleport/api/utils/retryutils/retry.go:184
main.main
	/go/src/github.com/gravitational/teleport/integrations/operator/main.go:125
runtime.main
	/opt/go/src/runtime/proc.go:250
2023-08-08T10:47:13Z	ERROR	setup	failed to setup teleport client	{"error": "context canceled"}
```

This is because we use the extraVolume to mount these in place for the
auth container.
When we enable the Operator sidecar, it fails to access these files.

This change fixes that by adding an extraVolumeMount to the operator
manifest.